Fix docker path on MacOS (#11)
When using the module on macos, the docker executable wasn't found,
resulting in a crash of the program.
I used the subprocess module to get the result of the shell command
`where docker` and use the result as path for builder.

I also renamed the folder "benchkit" in `/tmp` with "pythainer", I can
undo that if needed.
 
Tested on MacOS and Ubuntu 22.04.