Home sweet home (#18)
Enhance the `personnal_runner` to manage dotfiles (which are located in
`~/dotfiles/`) so as to have a container that looks like home.

The `preserve_history` argument is used to preserve shell history (eg:
bash history is stored in `.pythainer_history.bash` in the folder of the
python file that launches pythainer). Works with bash and zsh, but other
shells can easily be added.

Signed-off-by: Attilio <attilio.discepoli@vub.be>